+++ codex-rs/api-client/src/auth.rs
@@ -1,0 +1,15 @@
+use async_trait::async_trait;
+use codex_app_server_protocol::AuthMode;
+
+#[derive(Debug, Clone)]
+pub struct AuthContext {
+    pub mode: AuthMode,
+    pub bearer_token: Option<String>,
+    pub account_id: Option<String>,
+}
+
+#[async_trait]
+pub trait AuthProvider: Send + Sync {
+    async fn auth_context(&self) -> Option<AuthContext>;
+    async fn refresh_token(&self) -> Result<Option<String>, String>;
+}
