+++ codex-rs/api-client/src/lib.rs
@@ -1,8 +1,34 @@
+pub mod api;
+pub mod auth;
+pub mod chat;
+pub mod error;
+pub mod model_provider;
 pub mod prompt;
+pub mod responses;
 pub mod stream;
 
+pub use crate::api::ApiClient;
+pub use crate::auth::AuthContext;
+pub use crate::auth::AuthProvider;
+pub use crate::chat::AggregateStreamExt;
+pub use crate::chat::ChatAggregationMode;
+pub use crate::chat::ChatCompletionsApiClient;
+pub use crate::chat::ChatCompletionsApiClientConfig;
+pub use crate::error::Error;
+pub use crate::error::Result;
+pub use crate::model_provider::BUILT_IN_OSS_MODEL_PROVIDER_ID;
+pub use crate::model_provider::ModelProviderInfo;
+pub use crate::model_provider::WireApi;
+pub use crate::model_provider::built_in_model_providers;
+pub use crate::model_provider::create_oss_provider;
+pub use crate::model_provider::create_oss_provider_with_base_url;
 pub use crate::prompt::Prompt;
+pub use crate::responses::ResponsesApiClient;
+pub use crate::responses::ResponsesApiClientConfig;
+pub use crate::responses::stream_from_fixture;
 pub use crate::stream::Reasoning;
+pub use crate::stream::ResponseEvent;
+pub use crate::stream::ResponseStream;
 pub use crate::stream::TextControls;
 pub use crate::stream::TextFormat;
 pub use crate::stream::TextFormatType;
